requirements.txt
semantic-kernel
pyautogen
azure-cosmos
azure-identity
pandas
openai